Added the AudioData class, which contains a range of templated structrures for manipulating different sample type primitives. This will replace the old AudioDataConverters class, and I've refactored a lot of the audio devices and formats to use the new classes.

+++ src/native/mac/juce_mac_Fonts.mm
@@ -458,7 +458,7 @@ private:
             }
 
             // If we failed to find it "properly", this dodgy fall-back seems to do the trick for most fonts!
-            return jmax (-1, c - 29);
+            return jmax (-1, (int) c - 29);
         }
 
     private:
